requirements.txt
ccxt
python-telegram-bot
schedule
python-dotenv